Build: 934
Now, you can use all the blocks that are still in the SS, also, I can say that it will be possible to create your own blocks, but there is no way to add them to the SS (because it is not profitable for me), only through me and update how to draw up a file, I I'll show you later. The game part was implemented right now and I load projects in a different way. Bug from English solved